Removed test edit button
