+++ frontend/src/components/AdminDash/events/EventManager.jsx
@@ -5,7 +5,6 @@ import { Icon } from 'components/Shared';
 import EventTable from './EventTable';
 import { fetchEvents } from 'components/AdminDash/queries';
 import EventCreateModal from './EventCreateModal';
-import EventEditModal from './EventEditModal';
 
 const Styled = {
   Container: styled.div`
@@ -34,7 +33,6 @@ const EventManager = () => {
   const [loading, setLoading] = useState(true);
   const [events, setEvents] = useState([]);
   const [showCreateModal, setShowCreateModal] = useState(false);
-  const [showEditModal, setShowEditModal] = useState(false);
 
   const onRefresh = () => {
     setLoading(true);
@@ -49,20 +47,9 @@ const EventManager = () => {
       });
   };
 
-  const onEditClicked = () => {
-    setShowEditModal(true);
-  }
-
   const onCreateClicked = () => {
     setShowCreateModal(true);
   };
-
-  const toggleEditModal = () => {
-    setShowEditModal(prev => !prev);
-  };
-  useEffect(() => {
-    onRefresh();
-  }, []);
 
   const toggleCreateModal = () => {
     setShowCreateModal(prev => !prev);
@@ -82,14 +69,9 @@ const EventManager = () => {
           <Icon color="grey3" name="refresh" />
           <span> Refresh</span>
         </Styled.Button>
-        <Styled.Button onClick={onEditClicked}>
-          <Icon color="grey3" name="create" />
-          <span>Edit</span>
-        </Styled.Button>
       </Styled.HeaderContainer>
       <EventTable events={events} loading={loading}></EventTable>
       <EventCreateModal open={showCreateModal} toggle={toggleCreateModal} />
-      <EventEditModal open={showEditModal} toggle={toggleEditModal} />
     </Styled.Container>
   );
 };
